MDX init
