+++ package.json
@@ -5,6 +5,8 @@
   "dependencies": {
     "@fontsource/fira-code": "^4.2.1",
     "@loadable/component": "^5.6.1",
+    "@mdx-js/mdx": "^1.6.22",
+    "@mdx-js/react": "^1.6.22",
     "@tensorflow/tfjs": "^0.11.6",
     "animejs": "^2.2.0",
     "babel-plugin-styled-components": "^1.10.0",
@@ -16,6 +18,7 @@
     "gatsby-image": "^2.0.30",
     "gatsby-plugin-feed": "^2.0.11",
     "gatsby-plugin-google-analytics": "^2.0.8",
+    "gatsby-plugin-mdx": "^1.10.0",
     "gatsby-plugin-nprogress": "^2.0.7",
     "gatsby-plugin-offline": "^2.0.20",
     "gatsby-plugin-react-helmet": "^3.0.5",
@@ -30,7 +33,6 @@
     "gatsby-remark-smartypants": "^2.0.7",
     "gatsby-source-filesystem": "^2.0.12",
     "gatsby-transformer-javascript-frontmatter": "^2.0.5",
-    "gatsby-transformer-remark": "^2.1.18",
     "gatsby-transformer-sharp": "^2.1.10",
     "gray-matter": "^3.1.1",
     "loaders.css": "^0.1.2",
